Move all content into Nuxt Content
Previously, we had two pages which were directly rendered through Nuxt:
the homepage, and the news overview. This worked mostly fine, but caused
404 errors when Nuxt Content queried for `/news` and `/`. These 404
errors were harmless in (dynamic) production, but caused "nuxi generate"
to fail.

Move all content into Nuxt Content by simply making them templates. In
addition to resolving the 404 error, it also simplifies our directory
structure. Now all content is served through Nuxt Content in the
`content` directory -- simple and easy!
