+++ content/index.md
@@ -1,0 +1,7 @@
+---
+title: "FOSSi Foundation: the international not for profit organisation which promotes and protects the open source silicon chip movement"
+layout: default
+---
+
+::ff-homepage
+::
