+++ content/news.md
@@ -1,0 +1,8 @@
+---
+title: News
+layout: default
+header: News from the FOSSi Foundation
+---
+
+::ff-news-overview
+::
